requirements.txt
ggwave==0.4.0
pyaudio==0.2.13
transformers==4.36.2
torch==2.2.0
flask==3.0.2
python-dotenv==1.0.1
numpy==1.26.4
sounddevice==0.4.6
requests==2.31.0 